backend/requirements.txt
fastapi
uvicorn
python-multipart
chromadb
langchain
langchain-community
langgraph
ollama
pypdf
networkx
sentence-transformers
sqlalchemy
google-generativeai
python-dotenv
